Update README and Documentation
