--- src/test/java/Core.java
@@ -1,11 +1,0 @@
-import me.zero.alpine.bus.EventBus;
-import me.zero.alpine.bus.EventManager;
-
-public class Core {
-
-    public static final EventBus EVENT_BUS = EventManager.builder()
-        .setName("root")
-        .setRecursiveDiscovery()
-        .setSuperListeners()
-        .build();
-}
