--- src/test/java/EventProcessor.java
@@ -1,16 +1,0 @@
-import me.zero.alpine.listener.Subscribe;
-import me.zero.alpine.listener.EventSubscriber;
-import me.zero.alpine.listener.Listener;
-
-public class EventProcessor implements EventSubscriber {
-
-    @Subscribe
-    private Listener<String> stringListener = new Listener<>(s -> {
-        System.out.println("stringListener:" + s);
-    }, new LengthOf3Filter());
-
-    @Subscribe
-    private Listener<CharSequence> charSequenceListener = new Listener<>(s -> {
-        System.out.println("charSequenceListener:" + s);
-    }, new LengthOf3Filter());
-}
